requirements.txt
Django==5.0.12
gunicorn==21.2.0
azure-servicebus==7.12.1
python-dotenv==1.0.1
azure-core==1.38.0
mysqlclient==2.2.0
